Move bitstring into pkb/bitstring
