+++ pkg/bitstring/bitstring.go
