+++ pkg/bitstring/bitstring_test.go
