+++ pkg/bitstring/example_test.go
